Use the corrent license file and add MPL 2.0 headers
We inadvertedly left the EPL 1.0 license file that gets added by
default when you create a library using Leiningen, but the license of
the library is the MPL 2.0 (as stated in the README.md file).

So remove the EPL license file, add the correct MPL 2.0 license file
and add all the missing MPL 2.0 headers to all code files.